+++ src/Alert.js
@@ -1,6 +1,7 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
 import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
+import { mapToCssModules } from './utils';
 
 const FirstChild = ({ children }) => (
   React.Children.toArray(children)[0] || null
@@ -8,14 +9,15 @@ const FirstChild = ({ children }) => (
 
 const propTypes = {
   children: PropTypes.node,
-  className: PropTypes.any,
+  className: PropTypes.string,
+  cssModule: PropTypes.object,
   color: PropTypes.string,
   isOpen: PropTypes.bool,
   toggle: PropTypes.func,
   tag: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
   transitionAppearTimeout: PropTypes.number,
   transitionEnterTimeout: PropTypes.number,
-  transitionLeaveTimeout: PropTypes.number
+  transitionLeaveTimeout: PropTypes.number,
 };
 
 const defaultProps = {
@@ -30,6 +32,7 @@ const defaultProps = {
 const Alert = (props) => {
   const {
     className,
+    cssModule,
     tag: Tag,
     color,
     isOpen,
@@ -41,12 +44,12 @@ const Alert = (props) => {
     ...attributes
   } = props;
 
-  const classes = classNames(
+  const classes = mapToCssModules(classNames(
     className,
     'alert',
     `alert-${color}`,
     { 'alert-dismissible': toggle }
-  );
+  ), cssModule);
 
   const alert = (
     <Tag {...attributes} className={classes} role="alert">