+++ src/CardLink.js
@@ -1,9 +1,12 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
+import { mapToCssModules } from './utils';
 
 const propTypes = {
   tag: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
-  className: PropTypes.any
+  getRef: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
+  className: PropTypes.string,
+  cssModule: PropTypes.object,
 };
 
 const defaultProps = {
@@ -13,16 +16,18 @@ const defaultProps = {
 const CardLink = (props) => {
   const {
     className,
+    cssModule,
     tag: Tag,
+    getRef,
     ...attributes
   } = props;
-  const classes = classNames(
+  const classes = mapToCssModules(classNames(
     className,
     'card-link'
-  );
+  ), cssModule);
 
   return (
-    <Tag {...attributes} className={classes} />
+    <Tag {...attributes} ref={getRef} className={classes} />
   );
 };