+++ src/DropdownToggle.js
@@ -1,24 +1,26 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
+import { mapToCssModules } from './utils';
 import Button from './Button';
 
 const propTypes = {
   caret: PropTypes.bool,
   children: PropTypes.node,
-  className: PropTypes.any,
+  className: PropTypes.string,
+  cssModule: PropTypes.object,
   disabled: PropTypes.bool,
   onClick: PropTypes.func,
   'data-toggle': PropTypes.string,
   'aria-haspopup': PropTypes.bool,
   split: PropTypes.bool,
-  tag: PropTypes.oneOfType([PropTypes.func, PropTypes.string])
+  tag: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
+  nav: PropTypes.bool,
 };
 
 const defaultProps = {
   'data-toggle': 'dropdown',
   'aria-haspopup': true,
   color: 'secondary',
-  tag: Button
 };
 
 const contextTypes = {
@@ -39,6 +41,10 @@ class DropdownToggle extends React.Component {
       return;
     }
 
+    if (this.props.nav && !this.props.tag) {
+      e.preventDefault();
+    }
+
     if (this.props.onClick) {
       this.props.onClick(e);
     }
@@ -47,17 +53,29 @@ class DropdownToggle extends React.Component {
   }
 
   render() {
-    const { className, caret, split, tag: Tag, ...props } = this.props;
+    const { className, cssModule, caret, split, nav, tag, ...props } = this.props;
     const ariaLabel = props['aria-label'] || 'Toggle Dropdown';
-    const classes = classNames(
+    const classes = mapToCssModules(classNames(
       className,
       {
         'dropdown-toggle': caret || split,
         'dropdown-toggle-split': split,
         active: this.context.isOpen,
+        'nav-link': nav
       }
-    );
+    ), cssModule);
     const children = props.children || <span className="sr-only">{ariaLabel}</span>;
+
+    let Tag;
+
+    if (nav && !tag) {
+      Tag = 'a';
+      props.href = '#';
+    } else if (!tag) {
+      Tag = Button;
+    } else {
+      Tag = tag;
+    }
 
     return (
       <Tag