+++ src/InputGroupButton.js
@@ -1,13 +1,15 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
+import { mapToCssModules } from './utils';
 import Button from './Button';
 
 const propTypes = {
   tag: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
   children: PropTypes.node,
-  groupClassName: PropTypes.any,
+  groupClassName: PropTypes.string,
   groupAttributes: PropTypes.object,
-  className: PropTypes.any
+  className: PropTypes.string,
+  cssModule: PropTypes.object,
 };
 
 const defaultProps = {
@@ -17,6 +19,7 @@ const defaultProps = {
 const InputGroupButton = (props) => {
   let {
     className,
+    cssModule,
     tag: Tag,
     children,
     groupClassName,
@@ -25,10 +28,10 @@ const InputGroupButton = (props) => {
   } = props;
 
   if (typeof children === 'string') {
-    const groupClasses = classNames(
+    const groupClasses = mapToCssModules(classNames(
       groupClassName,
       'input-group-btn'
-    );
+    ), cssModule);
 
     return (
       <Tag {...groupAttributes} className={groupClasses}>
@@ -37,10 +40,10 @@ const InputGroupButton = (props) => {
     );
   }
 
-  const classes = classNames(
+  const classes = mapToCssModules(classNames(
     className,
     'input-group-btn'
-  );
+  ), cssModule);
 
   return (
     <Tag {...attributes} className={classes} children={children} />