+++ src/Modal.js
@@ -6,7 +6,8 @@ import Fade from './Fade';
 import {
   getOriginalBodyPadding,
   conditionallyUpdateScrollbar,
-  setScrollbarWidth
+  setScrollbarWidth,
+  mapToCssModules,
 } from './utils';
 
 const propTypes = {
@@ -21,7 +22,8 @@ const propTypes = {
   onEnter: PropTypes.func,
   onExit: PropTypes.func,
   children: PropTypes.node,
-  className: PropTypes.any
+  className: PropTypes.string,
+  cssModule: PropTypes.object,
 };
 
 const defaultProps = {
@@ -111,7 +113,7 @@ class Modal extends React.Component {
       this._element = null;
     }
 
-    document.body.className = classNames(classes).trim();
+    document.body.className = mapToCssModules(classNames(classes).trim(), this.props.cssModule);
     setScrollbarWidth(this.originalBodyPadding);
   }
 
@@ -123,16 +125,18 @@ class Modal extends React.Component {
     const classes = document.body.className;
     this._element = document.createElement('div');
     this._element.setAttribute('tabindex', '-1');
+    this._element.style.position = 'relative';
+    this._element.style.zIndex = '1000';
     this.originalBodyPadding = getOriginalBodyPadding();
 
     conditionallyUpdateScrollbar();
 
     document.body.appendChild(this._element);
 
-    document.body.className = classNames(
+    document.body.className = mapToCssModules(classNames(
       classes,
       'modal-open'
-    );
+    ), this.props.cssModule);
 
     this.renderIntoSubtree();
   }
@@ -169,9 +173,9 @@ class Modal extends React.Component {
             tabIndex="-1"
           >
             <div
-              className={classNames('modal-dialog', this.props.className, {
+              className={mapToCssModules(classNames('modal-dialog', this.props.className, {
                 [`modal-${this.props.size}`]: this.props.size
-              })}
+              }), this.props.cssModule)}
               role="document"
               ref={(c) => (this._dialog = c)}
             >