+++ webpack.dev.config.js
@@ -35,7 +35,11 @@ var paths = [
   '/components/tabs/',
   '/components/jumbotron/',
   '/components/alerts/',
+<<<<<<< HEAD
   '/components/listgroup/',
+=======
+  '/components/collapse/',
+>>>>>>> reactstrap/master
   '/404.html'
 ];
 
@@ -95,7 +99,7 @@ var config = [{
   resolve: {
     extensions: ['', '.js', '.json'],
     alias: {
-      'bootstrap-css': path.join(__dirname,'node_modules/bootstrap/dist/css/bootstrap.css'),
+      'bootstrap-css': path.join(__dirname,'node_modules/bootstrap/dist/css/bootstrap-flex.css'),
       reactstrap: path.resolve('./src')
     }
   }